chore: Move `album.utils.ts`
