+++ mobile/src/data/album/queries.ts
@@ -1,10 +1,10 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { useTranslation } from "react-i18next";
 
-import { AlbumArtistsKey } from "~/api/album.utils";
 import { Resynchronize } from "~/stores/Playback/actions";
 import { queries as q } from "~/queries/keyStore";
 import { updateAlbum } from "./api";
+import { AlbumArtistsKey } from "./utils";
 
 import { clearAllQueries } from "~/lib/react-query";
 import { formatSeconds } from "~/utils/number";
